Refactor spring-jms code samples
