+++ pom.xml
@@ -90,12 +90,13 @@
         <module>spring-hibernate3</module>
         <module>spring-hibernate4</module>
         <module>spring-jpa</module>
+        <module>spring-jms</module>
         <module>spring-katharsis</module>
         <module>spring-mockito</module>
         <module>spring-mvc-java</module>
         <module>spring-mvc-no-xml</module>
         <module>spring-mvc-xml</module>
-	<module>spring-mvc-tiles</module>
+        <module>spring-mvc-tiles</module>
         <module>spring-openid</module>
         <module>spring-protobuf</module>
         <module>spring-quartz</module>
