--- spring-jms/src/com/baeldung/spring/jms/SampleJMSExample.java
@@ -1,17 +1,0 @@
-package com.baeldung.spring.jms;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-
-import org.apache.activemq.broker.BrokerFactory;
-import org.apache.activemq.broker.BrokerService;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-public class SampleJMSExample {
-    public static void main(String[] args) throws URISyntaxException, Exception {
-        BrokerService broker = BrokerFactory.createBroker(new URI("broker:(tcp://localhost:61616)"));
-        broker.start();
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
-
-    }
-}
