+++ spring-jms/src/main/java/com/baeldung/spring/jms/Employee.java
@@ -1,0 +1,23 @@
+package com.baeldung.spring.jms;
+
+public class Employee {
+    private String name;
+    private Integer age;
+
+    public Employee(String name, Integer age) {
+        this.name = name;
+        this.age = age;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Integer getAge() {
+        return age;
+    }
+
+    public String toString() {
+        return "Person: name(" + name + "), age(" + age + ")";
+    }
+}
