+++ spring-jms/src/main/java/com/baeldung/spring/jms/SampleJMSExample.java
@@ -1,0 +1,16 @@
+package com.baeldung.spring.jms;
+
+import org.apache.activemq.broker.BrokerFactory;
+import org.apache.activemq.broker.BrokerService;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+import java.net.URI;
+
+public class SampleJMSExample {
+    public static void main(String[] args) throws Exception {
+        BrokerService broker = BrokerFactory.createBroker(new URI("broker:(tcp://localhost:61616)"));
+        broker.start();
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("applicationContext.xml");
+
+    }
+}
