+++ spring-jms/src/main/java/com/baeldung/spring/jms/SampleJmsMessageSender.java
@@ -1,0 +1,28 @@
+package com.baeldung.spring.jms;
+
+import org.springframework.jms.core.JmsTemplate;
+import org.springframework.jms.core.MessageCreator;
+
+import javax.jms.*;
+
+public class SampleJmsMessageSender {
+
+    private JmsTemplate jmsTemplate;
+    private Queue queue;
+
+    public void createJmsTemplate(ConnectionFactory cf) {
+        this.jmsTemplate = new JmsTemplate(cf);
+    }
+
+    public void setQueue(Queue queue) {
+        this.queue = queue;
+    }
+
+    public void simpleSend() {
+        this.jmsTemplate.send(this.queue, new MessageCreator() {
+            public Message createMessage(Session session) throws JMSException {
+                return session.createTextMessage("hello queue world");
+            }
+        });
+    }
+}
